libdmabufinfo: Rename show_dmabuf_sysfs_stats to show_per_buffer_stats
The '-b' argument to dmabuf_dump is for:

"-b	 show DMA-BUF per-buffer, per-exporter and per-device statistics"

We are adding the ability to obtain per-buffer dmabuf information from
BPF as a replacement for sysfs. Rename the internal variable to match
the "per-buffer" language for the flag instead of specifying the
interface used to get it.

Bug: 442596155
Test: atest dmabufinfo_test
Flag: PURE_REFACTOR
Change-Id: Icf30b364c9d1d49683956240ce37cf3fc967b068