libdmabufinfo: Add BPF per-buffer stats
BPF is an alternative (and more efficient) source of per-dmabuf
information that is currently obtained from the deprecated [1] sysfs
interface [2].

Move callers off of the specific sysfs named function
GetDmabufSysfsStats(), and onto a generic GetDmabufPerBufferStats()
which tries BPF first and then falls back to sysfs. As part of this
refactor, callers must provide already-collected per-buffer data to the
libdmabufinfo API.

As a result of this change, ReadBufferExporter(), ReadBufferSize(),
and GetDmabufTotalExportedKb() become unused, and have been removed.
(Their symbols are retained for GRF compatibility, but no new uses are
permitted.)

[1] https://lore.kernel.org/all/20220617085702.4298-1-christian.koenig@amd.com/
[2] https://lore.kernel.org/all/20210603214758.2955251-1-hridya@google.com/

Bug: 442596155
Test: atest dmabufinfo_test
Test: atest libmeminfo_test
Change-Id: Id9c6e91b55f2d5763164851d34f86d4b5d5d2d6b