wsstop: Add option to only show private mappings
Idle page tracking tells us whether a page was accessed since the last
check, but not by which process(es). This means that for shared pages
we cannot determine if they're actually part of the specific process'
working set, or were accessed only by unrelated processes.

Add a new -p flag that filters out VMAs that only have shared pages,
showing only a subset of mappings guaranteed to be part of the working
set.

Test: wsstop -p -n 1 <pid>
Change-Id: I60ecfc2ca47b692c46d095377a2720d5275be397