libelfutils: Introduce generic ElfFile class
Rename Elf64_Sc to Elf_Sc to make it suitable for use with
both 32 and 64-bit ELFs.

Introduce a generic ElfFile class to handle both ELF32 and ELF64
files. This replaces the Elf64Binary class and moves the
implementation to a new elf-file.cpp file.

The ElfFile class serves as an interface for ELF file operations.
The actual implementation is provided by the templated ElfFileImpl
class, which is specialized for ELF32 and ELF64 types. This design
abstracts the necessary template details from the callers.

A factory method, ElfFile::create(), detects the ELF class (32 or
64-bit) and returns a std::unique_ptr<ElfFile>, providing a
clean and type-safe way for clients to interact with ELF files
without needing to know the underlying implementation details.

Test: atest vts_built_with_ddk_test
Test: atest Vts16KPageSizeTest
Bug: 433965396
Bug: 433514658
Change-Id: Ic1532539712df895af92e4747fa5c1f4c514e13b
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>