libelfutils: Refactor ElfComparator for 32/64-bit ELF comparison
Refactor the Elf64Comparator to support comparing both 32-bit and
64-bit ELF files.

Replace the static comparison methods with a generic ElfComparator
interface and a templated implementation. This aligns the comparator
with the design of the ElfFile, ElfParser, and ElfWriter classes.

The new design ensures that comparisons are only performed between
ELF files of the same architecture. A factory function,
ElfComparator::create(), returns a nullptr if the two ELF files being
compared do not have the same class, preventing invalid comparisons.

Test: atest vts_built_with_ddk_test
Test: atest Vts16KPageSizeTest
Bug: 433965396
Bug: 433514658
Change-Id: Ia5fdabe490f269951c0b0ab459dc8184177a4b2f
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>