libelfutils: Relocate gen_invalid_elf and add 32/64-bit support
Relocate the gen_invalid_libs tool to tools/gen_invalid_elf and
refactor it to support both 32-bit and 64-bit ELF files.

The logic for creating invalid ELF files has been moved into a set of
public methods on the ElfFileImpl class. This allows the class to
maintain control over its own state while providing a clear API for
this specialized testing tool.

Introduce parameters to override the offsets at which the ElfWriter
writers the various pieces of the ELF. This is used by the
gen_invalid_libs tools to write corrupted ELFs headers without
risk of overwriting other parts of the ELF.

The new gen_invalid_elf binary uses these methods to create malformed
ELF files for testing purposes.

The command-line interface is updated to be more explicit, using
flags to specify the reference ELF and the type of invalid file to
generate.

Test: atest bionic-unit-test (with new gen_invalid_elf genrules)
Test: atest vts_built_with_ddk_test
Test: atest Vts16KPageSizeTest
Bug: 433965396
Bug: 433514658
Change-Id: I79d945824d7637859ad7af041af847e75427fdb0
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>