libelfutils: Relocate compare_libs and add 32/64-bit support
Relocate the compare_libs binary to tools/compare_elfs and
refactor it to support both 32-bit and 64-bit ELF files.

Update the compare_elfs tool to use the generic ElfFile and
ElfComparator interfaces, allowing it to compare ELF files
of any supported architecture.

Align the comparison tool with the new, refactored architecture of
the library and improve the overall project structure.

Example output:

    ./compare_elfs libtest_empty.so libtest_invalid-rw_load_segment.so

    Comparing ELF Executable Headers ...
    Executable Headers are equal

    Comparing ELF Program Headers ...
            Diff phdrs1[2].p_flags = 0x5 != phdrs2[2].p_flags = 0x7
    Program Headers are NOT equal

    Comparing ELF Section Headers ...
    Section Headers are equal

    Comparing ELF Sections (content) ...
    Sections are equal

Test: ./compare_elfs libtest_empty.so libtest_invalid-rw_load_segment.so
Bug: 433965396
Bug: 433514658
Change-Id: I93d08267ececa7c0ab3948b923eadc293f87196c
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>