elf_alignment_test: Add 16K RELRO validation checks
Add validation for the PT_GNU_RELRO segment in the elf_alignment_test
to ensure it can be correctly loaded on 16KB devices.

If the RELRO is the entire LOAD segment then alignment is guaranteed
by max-page-size. This is verified by checking that they both contain
the same sections.

If the RELRO start vaddr is not aligned it must be the at the start
of the corresponding LOAD segment (i.e it's the prefix).

If the RELRO end vaddr is not aligned it must be the at the end
of the corresponding LOAD segment (i.e it's the suffix).

The test failure messages have also been improved to include actual
and expected alignment values in hexadecimal format for easier debugging.

Bug: 451068747
Bug: 448550643
Test: atest elf_alignment_test
Change-Id: I4f4e21e39f289685cbb65e2b8facb89e6197639c
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>