libelfutils: Add RELRO end address alignment check
Add a function to get the end address of the PT_GNU_RELRO segment and
verify its alignment in the elf_alignment_test. The RELRO segment needs
to be properly aligned to the maximum supported page size to ensure
correct memory protection.

Also, improve the test failure messages to include the actual and
expected alignment values in hexadecimal format for easier debugging.

Bug: 448550643
Test: atest elf_alignment_test
Change-Id: I0f021fcb71bc831d70d9d4313f2996c6157c2545
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>