libdmabufinfo: VTS: Require BPF dmabuf iterator on kernels >= 6.18
The 6.18 GKI release will be the first without CONFIG_DMABUF_SYSFS_STATS
which used to be the only source of per-dmabuf information before commit
fa00c56 ("libdmabufinfo: Add BPF per-buffer stats"), so the dmabuf
iterator is required to be functional for kernels 6.18 or newer.

While the dmabuf iterator has been backported to all but one older GKI
kernel (android12-5.10: b/445433934), those older kernels still have
CONFIG_DMABUF_SYSFS_STATS and the platform will fallback to sysfs when
BPF is not usable, so BPF is an optional performance enhancement for
devices running those kernels.

Test: vts_dmabufinfo_test
Bug: 454669973
Bug: 455593147
Change-Id: I2f1ed1f8340ab6d0cff41ac5b910d5fa5d581f37