libelfutils: Refactor ElfParser to simplify internal implementation
Refactor the ElfParser class to operate directly on the ElfFile
object it is parsing, rather than passing members as arguments to its
internal methods. This can be achieved sincethe ElfParser class is
friended within ElfFileImpl, granting it direct access to the
ElfFile's internal structures.

This change simplifies the parsing logic, reduces code duplication,
and improves the overall readability and maintainability of the
parser. It also serves as a prerequisite for enabling more granular,
robust parsing by downstream tools.

Bug: 454646441
Test: atest elf_alignment_test
Change-Id: I6ab6c6d36a49ec07b25e0e34495d0bde1c1e986e
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>