libelfutils: Add findSectionByName method to ElfFile
Introduce a new findSectionByName() method to the base ElfFile class
to allow clients to search for a section by its name.

The implementation resides in the base class and uses the virtual
getSections() method to perform the search. This design avoids code
duplication across the 32-bit and 64-bit ElfFile implementations
and centralizes the search logic.

Bug: 454646441
Test: m libelfutils
Change-Id: Ib46202a8d73c02cf9bc7a042f1a35ea16bbadc0b
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>